change long to int + WIP

+++ src/main/java/Clock.java
@@ -1,9 +1,9 @@
 package main.java;
 
 public class Clock {
-    private long time;
+    private int time;
 
-    public long getTime() {
+    public int getTime() {
         return time;
     }
 }
